Fixed bug in rotation function, changed randomizatioon of the direction, downgraded the speed when duck is free

+++ src/tema1/Bullets.hpp
@@ -57,7 +57,7 @@ public:
 			m_bullets.emplace_back(new Bullet({position, 0, 0}));
 		}
 	}
-	int GetRemainingBullets() { return m_bullets.size(); };
+	int GetRemainingBullets() { return static_cast<int>(m_bullets.size()); };
 
 private:
 	std::vector<std::unique_ptr<Bullet>> m_bullets;
